+++ packages/app/src/renderer/src/components/Artplayer/Index.vue
@@ -20,6 +20,7 @@ const props = withDefaults(
     option: {
       fullscreen?: boolean;
       url?: string;
+      type?: string;
       string?: any;
       plugins?: {
         heatmap?: {
@@ -142,7 +143,8 @@ onMounted(async () => {
     return;
   }
   instance = new Artplayer({
-    url: "",
+    url: props?.option?.url ?? "",
+    type: props?.option?.type ?? "",
     isLive: props.isLive,
     ...props.option,
     container: artRef.value,
@@ -166,7 +168,6 @@ onMounted(async () => {
           });
           flv.attachMediaElement(video);
           flv.load();
-          flv.play();
 
           art.flv = flv;
           art.on("destroy", () => flv.destroy());